+++ lib/eml/language/native.ex
@@ -21,13 +21,5 @@ defmodule Eml.Language.Native do
     Float.to_string(data)
   end
 
-  def parse({ :escaped, string }, Tuple) do
-    { :escaped, string }
-  end
-  def parse(data, Tuple) do
-    { :error, "Unparsable data: #{inspect data}" }
-  end
-
-
   def render(data, _opts), do: data
 end