+++ test/eml_test.exs
@@ -1,37 +1,35 @@
 defmodule EmlTest do
   use ExUnit.Case
-  use Eml
+  use Eml.Language.Html
   alias Eml.Element, as: M
 
   defp doc() do
-    [%M{tag: :html, content: [
-      %M{tag: :head, attrs: %{class: "test"}, content: [
-        %M{tag: :title, attrs: %{class: "title"}, content: ["Eml is Html for developers"]}
-      ]},
-      %M{tag: :body, attrs: %{class: ["test", "main"]}, content: [
-        %M{tag: :h1, attrs: %{class: "title"}, content: ["Eml is Html for developers"]},
-        %M{tag: :article, attrs: %{class: "content", "data-custom": "some custom attribute"},
-           content: ["Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."]},
-        %M{tag: :div, attrs: %{id: "main-side-bar", class: ["content", "side-bar"]}, content: [
-          %M{tag: :span, attrs: %{class: "test"}, content: ["Some notes..."]}
-        ]}
-      ]}
-    ]}]
+    %M{tag: :html, content: [
+     %M{tag: :head, attrs: %{class: "test"}, content: [
+       %M{tag: :title, attrs: %{class: "title"}, content: ["Eml is Html for developers"]}
+     ]},
+     %M{tag: :body, attrs: %{class: ["test", "main"]}, content: [
+       %M{tag: :h1, attrs: %{class: "title"}, content: ["Eml is Html for developers"]},
+       %M{tag: :article, attrs: %{class: "content", "data-custom": "some custom attribute"},
+          content: ["Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."]},
+       %M{tag: :div, attrs: %{id: "main-side-bar", class: ["content", "side-bar"]}, content: [
+         %M{tag: :span, attrs: %{class: "test"}, content: ["Some notes..."]}
+       ]}
+     ]}
+   ]}
   end
 
   test "Element macro" do
-    doc = eml do
-      html do
-        head class: "test" do
-          title [class: "title"], "Eml is Html for developers"
-        end
-        body class: ["test", "main"] do
-          h1 [class: "title"], "Eml is Html for developers"
-          article [class: "content", "data-custom": "some custom attribute"],
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
-          div id: "main-side-bar", class: ["content", "side-bar"] do
-            span [class: "test"], "Some notes..."
-          end
+    doc = html do
+      head class: "test" do
+        title [class: "title"], "Eml is Html for developers"
+      end
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
+        article [class: "content", "data-custom": "some custom attribute"],
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
+        div id: "main-side-bar", class: ["content", "side-bar"] do
+          span [class: "test"], "Some notes..."
         end
       end
     end
@@ -40,38 +38,31 @@ defmodule EmlTest do
   end
 
   test "Element macro as match pattern" do
-    e = eml do
-      span(%{id: id}, _) = %M{tag: :span, attrs: %{id: "test"}, content: []}
-      id
-    end
-    assert "test" == unpack(e)
+    span(%{id: id}, _) = %M{tag: :span, attrs: %{id: "test"}, content: []}
+    assert "test" == id
   end
 
   test "Enumerate content" do
-    assert true == Enum.member?(unpack(doc()), "Some notes...")
+    assert true == Enum.member?(doc(), "Some notes...")
 
-    e = eml do
-      h1 [class: "title"], ["Eml is Html for developers"]
-    end
-    assert e == Enum.filter(unpack(doc()), &Element.has?(&1, tag: :h1))
+    e = [h1([class: "title"], "Eml is Html for developers")]
+    assert e == Enum.filter(doc(), &Eml.Element.has?(&1, tag: :h1))
 
-    e = eml do
-      [
-        head class: "test" do
-          title [class: "title"], "Eml is Html for developers"
-        end,
-        body class: ["test", "main"] do
-          h1 [class: "title"], "Eml is Html for developers"
-          article [class: "content", "data-custom": "some custom attribute"],
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
-          div id: "main-side-bar", class: ["content", "side-bar"] do
-            span [class: "test"], "Some notes..."
-          end
-        end,
-        span([class: "test"], "Some notes...")
-      ]
-    end
-    assert e == Enum.filter(unpack(doc()), &Element.has?(&1, class: "test"))
+    e = [
+      head class: "test" do
+        title [class: "title"], "Eml is Html for developers"
+      end,
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
+        article [class: "content", "data-custom": "some custom attribute"],
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
+        div id: "main-side-bar", class: ["content", "side-bar"] do
+          span([class: "test"], "Some notes...")
+        end
+      end,
+      span([class: "test"], "Some notes...")
+    ]
+    assert e == Enum.filter(doc(), &Eml.Element.has?(&1, class: "test"))
   end
 
   test "Render => Parse => Compare" do
@@ -79,100 +70,82 @@ defmodule EmlTest do
   end
 
   test "Types" do
-    assert :content   == Eml.type eml do: div([], 42)
-    assert :content   == Eml.type eml do: [1,2,"z"]
-    assert :content   == Eml.type eml do: ""
-    assert :content   == Eml.type eml do: []
-    assert :content   == Eml.type eml do: nil
-    assert :element    == Eml.type Eml.Element.new()
-    assert :element    == Eml.type unpack(eml do: div([], 42))
+    assert :element   == Eml.type div(42)
+    assert :element   == Eml.type Eml.Element.new()
     assert :binary    == Eml.type "strings are binaries"
-    assert :binary    == Eml.type Eml.unpackr(eml do: div([], 42))
-    assert :binary    == Eml.type unpack(eml do: [1,2,"z"])
-    assert :binary    == Eml.type Eml.render!((eml do: :name), [], render_params: true)
-    assert :binary    == Eml.type (eml do: :name)
+    assert :binary    == Eml.type Eml.unpackr div(42)
+    assert :binary    == Eml.type Eml.unpack Eml.parse!([1,2,"z"], Eml.Language.Native)
+    assert :binary    == Eml.type Eml.render!(Eml.parse!(:a, Eml.Language.Native), [], render_params: true)
+    assert :binary    == Eml.type Eml.parse!(:name, Eml.Language.Native)
                                   |> Eml.compile!()
                                   |> Eml.render!(name: "Vincent")
-    assert :template  == Eml.type Eml.compile!(eml do: :name)
-    assert :template  == Eml.type Eml.compile!(eml do: [div([], 1), div([], 2), div([], :param), "..."])
+    assert :template  == Eml.type Eml.compile!(Eml.parse!(:name, Eml.Language.Native))
+    assert :template  == Eml.type Eml.compile!([div([], 1), div([], 2), div([], :param), "..."])
     assert :parameter == Eml.type %Eml.Parameter{}
-    assert :parameter == Eml.type unpack(eml do: :param)
-  end
-
-  test "Native" do
-    assert []                  == eml do: [nil, "", []]
-    assert ["truefalse"]       == eml do: [true, false]
-    assert ["12345678"]        == eml do: Enum.to_list(1..8)
-    assert ["Hello world"]     == eml do: ["H", ["el", "lo", [" "]], ["wor", ["ld"]]]
-    assert ["Happy new 2015!"] == eml do: ["Happy new ", 2, 0, 1, 5, "!"]
   end
 
   test "Native parse" do
-    assert ["1234"]                          == Eml.parse([1,2,3,4], Eml.Language.Native)
+    assert []                  == Eml.parse! [nil, "", []], Eml.Language.Native
+    assert ["truefalse"]       == Eml.parse! [true, false], Eml.Language.Native
+    assert ["12345678"]        == Eml.parse! Enum.to_list(1..8), Eml.Language.Native
+    assert ["Hello world"]     == Eml.parse! ["H", ["el", "lo", [" "]], ["wor", ["ld"]]], Eml.Language.Native
+    assert ["Happy new 2015!"] == Eml.parse! ["Happy new ", 2, 0, 1, 5, "!"], Eml.Language.Native
     assert { :error, "Unparsable data: {}" } == Eml.parse({}, Eml.Language.Native)
   end
 
   test "Unpack" do
-    e = eml do: div([], 42)
-    assert "42" == unpack unpack e
-    assert "42" == unpack ["42"]
-    assert "42" == unpack "42"
-    assert "42" == unpack Eml.Element.new(:div, [], 42)
+    e = div 42
+    assert "42" == Eml.unpack e
+    assert "42" == Eml.unpack ["42"]
+    assert "42" == Eml.unpack "42"
+    assert "42" == Eml.unpack Eml.Element.new(:div, [], 42)
 
-    e = eml do: [div([], 1), div([], 2)]
-    assert e == unpack e
+    e = [div(1), div(2)]
+    assert e == Eml.unpack e
   end
 
   test "Single unpackr" do
-    single = eml do
-      html do
-        body do
-          div [], 42
-        end
+    single = html do
+      body do
+        div 42
       end
     end
     assert "42" == Eml.unpackr(single)
   end
 
   test "Multi unpackr" do
-    multi = eml do
-      html do
-        body do
-          div [], [ span([], 1), span([], 2) ]
-          div [], span([], 3)
-        end
+    multi = html do
+      body do
+        div [ span(1), span(2) ]
+        div span(3)
       end
     end
     assert [["1", "2"], "3"] == Eml.unpackr(multi)
   end
 
   test "Funpackr" do
-    multi = eml do
-      html do
-        body do
-          div [], [span([], 1), span([], 2)]
-          div [], span([], 3)
-        end
+    multi = html do
+      body do
+        div [span(1), span(2)]
+        div span(3)
       end
     end
     assert ["1", "2", "3"] == Eml.funpackr(multi)
   end
 
   test "Add content" do
-    input = eml do: body(id: "test")
-    to_add = eml do: div([], "Hello world!")
-    expected = eml do: body([id: "test"], to_add)
+    input = body(id: "test")
+    to_add = div("Hello world!")
+    expected = body([id: "test"], to_add)
 
     assert expected == Eml.add(input, to_add, id: "test")
   end
 
   test "Select by class1" do
-    expected = eml do
-      [
-       title([class: "title"], "Eml is Html for developers"),
-       h1([class: "title"], "Eml is Html for developers")
-      ]
-    end
+    expected = [
+      title([class: "title"], "Eml is Html for developers"),
+      h1([class: "title"], "Eml is Html for developers")
+    ]
     result = Eml.select(doc(), class: "title")
 
     # The order of the returned content is unspecified,
@@ -182,42 +155,36 @@ defmodule EmlTest do
   end
 
   test "Select by class 2" do
-    expected = eml do
-      [
-       head([class: "test"], title([class: "title"], "Eml is Html for developers")),
-       body class: ["test", "main"] do
-         h1 [class: "title"], "Eml is Html for developers"
-         article [class: "content", "data-custom": "some custom attribute"],
-         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
-         div id: "main-side-bar", class: ["content", "side-bar"] do
-           span [class: "test"], "Some notes..."
-         end
-       end,
-       span([class: "test"], "Some notes...")
-      ]
-    end
+    expected = [
+      head([class: "test"], title([class: "title"], "Eml is Html for developers")),
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
+        article [class: "content", "data-custom": "some custom attribute"],
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
+        div id: "main-side-bar", class: ["content", "side-bar"] do
+          span [class: "test"], "Some notes..."
+        end
+      end,
+      span([class: "test"], "Some notes...")
+    ]
     result = Eml.select(doc(), class: "test")
 
     assert Enum.all?(result, fn node -> node in expected end)
   end
 
   test "Select by id" do
-    expected = eml do
-      div id: "main-side-bar", class: ["content", "side-bar"] do
-        span [class: "test"], "Some notes..."
-      end
-    end
+    expected = [div id: "main-side-bar", class: ["content", "side-bar"] do
+      span [class: "test"], "Some notes..."
+    end]
     result = Eml.select(doc(), id: "main-side-bar")
 
     assert expected == result
   end
 
   test "Select by id and class 1" do
-    expected = eml do
-      div id: "main-side-bar", class: ["content", "side-bar"] do
-        span [class: "test"], "Some notes..."
-      end
-    end
+    expected = [div id: "main-side-bar", class: ["content", "side-bar"] do
+      span [class: "test"], "Some notes..."
+    end]
 
     result = Eml.select(doc(), id: "main-side-bar", class: "content")
 
@@ -236,14 +203,12 @@ defmodule EmlTest do
   end
 
   test "Remove by class 1" do
-    expected = eml do
-      html do
-        head [class: "test"] do
-          title [class: "title"], "Eml is Html for developers"
-        end
-        body class: ["test", "main"] do
-          h1 [class: "title"], "Eml is Html for developers"
-        end
+    expected = html do
+      head [class: "test"] do
+        title [class: "title"], "Eml is Html for developers"
+      end
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
       end
     end
     result = Eml.remove(doc(), class: "content")
@@ -252,21 +217,19 @@ defmodule EmlTest do
   end
 
   test "Remove by class 2" do
-    expected = eml do: html([], [])
+    expected = html []
     result   = Eml.remove(doc(), class: "test")
 
     assert expected == result
   end
 
   test "Remove by id" do
-    expected = eml do
-      html do
-        head [class: "test"], do: title([class: "title"], "Eml is Html for developers")
-        body class: ["test", "main"] do
-          h1 [class: "title"], "Eml is Html for developers"
-          article [class: "content", "data-custom": "some custom attribute"],
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
-        end
+    expected = html do
+      head [class: "test"], do: title([class: "title"], "Eml is Html for developers")
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
+        article [class: "content", "data-custom": "some custom attribute"],
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
       end
     end
     result = Eml.remove(doc(), id: "main-side-bar")
@@ -275,14 +238,12 @@ defmodule EmlTest do
   end
 
   test "Remove by id and class" do
-    expected = eml do
-      html do
-        head [class: "test"], do: title([class: "title"], "Eml is Html for developers")
-        body class: ["test", "main"] do
-          h1 [class: "title"], "Eml is Html for developers"
-          article [class: "content", "data-custom": "some custom attribute"],
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
-        end
+    expected = html do
+      head [class: "test"], do: title([class: "title"], "Eml is Html for developers")
+      body class: ["test", "main"] do
+        h1 [class: "title"], "Eml is Html for developers"
+        article [class: "content", "data-custom": "some custom attribute"],
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam suscipit non neque pharetra dignissim."
       end
     end
     result = Eml.remove(doc(), id: "main-side-bar", class: "content")
@@ -296,61 +257,55 @@ defmodule EmlTest do
 
   test "Parameters" do
     param = %Eml.Parameter{id: :a_parameter}
-    assert [param]               == eml do: :a_parameter
-    assert [param, "and", param] == eml do: [:a_parameter, "and", :a_parameter]
+    assert param                 == Eml.parse!(:a_parameter, Eml.Language.Native)
+    assert [param, "and", param] == Eml.parse!([:a_parameter, "and", :a_parameter], Eml.Language.Native)
   end
 
   test "Templates 1" do
-    e = eml do
-      div id: :myid do
-        div [], :fruit1
-        div [], :fruit2
-      end
+    e = div id: :myid do
+      div :fruit1
+      div :fruit2
     end
     { :ok, t } = Eml.compile(e)
 
     assert :template == Eml.type t
-    assert false == Template.bound?(t)
-    assert [:myid, :fruit1, :fruit2] == Template.unbound(t)
+    assert false == Eml.Template.bound?(t)
+    assert [:myid, :fruit1, :fruit2] == Eml.Template.unbound(t)
 
-    t = Template.bind(t, :myid, "double-fruit")
-    assert [:fruit1, :fruit2] == Template.unbound(t)
+    t = Eml.Template.bind(t, :myid, "double-fruit")
+    assert [:fruit1, :fruit2] == Eml.Template.unbound(t)
 
-    t = Template.bind(t, fruit1: "orange", fruit2: "lemon")
-    assert Template.bound?(t)
+    t = Eml.Template.bind(t, fruit1: "orange", fruit2: "lemon")
+    assert Eml.Template.bound?(t)
   end
 
   test "Templates 2" do
-    e = eml do
-      for _ <- 1..4 do
-        div [], :fruit
-      end
+    e = for _ <- 1..4 do
+      div [], :fruit
     end
     { :ok, t } = Eml.compile(e)
 
     assert :template == Eml.type t
-    assert false == Template.bound?(t)
-    assert [:fruit] == Template.unbound(t)
+    assert false == Eml.Template.bound?(t)
+    assert [:fruit] == Eml.Template.unbound(t)
 
-    t = Template.bind(t, fruit: "lemon")
-    assert [] == Template.unbound(t)
+    t = Eml.Template.bind(t, fruit: "lemon")
+    assert [] == Eml.Template.unbound(t)
 
     assert "<div>lemon</div><div>lemon</div><div>lemon</div><div>lemon</div>" ==
       Eml.render!(t)
   end
 
   test "Templates in eml" do
-    fruit  = eml do: section([], :fruit)
+    fruit  = section :fruit
     tfruit = Eml.compile!(fruit)
-    aside  = eml do: aside([], tfruit)
+    aside  = aside tfruit
     taside = Eml.compile!(aside)
 
     assert :template == Eml.type taside
 
-    expected = eml do
-      aside do
-        section [], "lemon"
-      end
+    expected = aside do
+      section "lemon"
     end
 
     assert Eml.render!(expected) == Eml.render!(taside, fruit: "lemon")
@@ -360,18 +315,16 @@ defmodule EmlTest do
     html       = "<div><span>#param{name}</span><span>#param{age}</span></div"
     name_param = %Eml.Parameter{id: :name}
     age_param  = %Eml.Parameter{id: :age}
-    eml        = Eml.parse(html)
+    eml        = Eml.parse!(html)
 
     assert [name_param, age_param] == Eml.unpackr eml
   end
 
   test "Parameterized attribute rendering" do
-    e = eml do
-      div id: :id_param,
-          class: [:class1, "class2", :class3],
-          _custom1: :custom,
-          _custom2: :custom
-    end
+    e = div id: :id_param,
+            class: [:class1, "class2", :class3],
+            _custom1: :custom,
+            _custom2: :custom
 
     expected1 = "<div data-custom1='#param{custom}' data-custom2='#param{custom}' class='#param{class1} class2 #param{class3}' id='#param{id_param}'></div>"
     assert expected1 == Eml.render!(e, [], render_params: true)
@@ -382,9 +335,9 @@ defmodule EmlTest do
                                        class3: "class3",
                                        custom: 1)
 
-    { :ok, t } = Eml.compile(e, [class3: "class3", custom: 1])
+    { :ok, t } = Eml.compile(e, class3: "class3", custom: 1)
     assert :template == Eml.type t
-    assert Enum.sort([:id_param, :class1]) == Enum.sort(Template.unbound(t))
+    assert Enum.sort([:id_param, :class1]) == Enum.sort(Eml.Template.unbound(t))
     assert expected2 == Eml.render!(t, id_param: "parameterized", class1: "class1")
   end
 end